feat: event 정보 삭제 API 추가, eventId int -> long으로 변경

+++ src/main/java/snowball/lesson/dto/GetLessonDetailsDto.java
@@ -22,7 +22,7 @@ public class GetLessonDetailsDto {
     private String content1;
     private String content2;
     private String thumbnail;
-    private int eventId;
+    private long eventId;
     private int discountRate;
     private LocalDateTime discountStartDate;
     private LocalDateTime discountFinishDate;
